change bins to 32 bits

+++ include/sysemu/balloon.h
@@ -19,9 +19,9 @@
 
 typedef void (QEMUBalloonEvent)(void *opaque, ram_addr_t target);
 typedef void (QEMUBalloonWorkingSetRequest)(void *opaque);
-typedef void (QEMUBalloonWorkingSetConfig)(void *opaque, uint64_t i0,
-                                  uint64_t i1, uint64_t i2, uint64_t refresh,
-                                  uint64_t report);
+typedef void (QEMUBalloonWorkingSetConfig)(void *opaque, uint32_t i0,
+                                  uint32_t i1, uint32_t i2, uint32_t refresh,
+                                  uint32_t report);
 typedef void (QEMUBalloonStatus)(void *opaque, BalloonInfo *info);
 
 int qemu_add_balloon_handler(QEMUBalloonEvent *event_func,
